Removed old IAGSStream as turned to be unused and redundant
