+++ Common/CMakeLists.txt
@@ -134,7 +134,6 @@ target_sources(common
     util/filestream.h
     util/geometry.cpp
     util/geometry.h
-    util/iagsstream.h
     util/ini_util.cpp
     util/ini_util.h
     util/inifile.cpp
