+++ Common/util/stream.h
@@ -14,7 +14,11 @@
 //
 // Base stream class.
 //
-// Provides default implementation for a few helper methods.
+// Provides a stream interface and a few helper methods.
+//
+// The user is advised to use advanced helper methods, such as Read/WriteX
+// and Read/WriteArrayOfX to allow the stream implementation properly control
+// endianness conversions when needed.
 // 
 // Only streams with uncommon behavior should be derived directly from Stream.
 // Most I/O devices should inherit DataStream instead.
@@ -23,7 +27,6 @@
 #ifndef __AGS_CN_UTIL__STREAM_H
 #define __AGS_CN_UTIL__STREAM_H
 
-#include "util/iagsstream.h"
 #include "util/string.h"
 
 namespace AGS
@@ -38,42 +41,100 @@ enum StreamWorkMode
     kStream_Write
 };
 
+enum StreamSeek
+{
+    kSeekBegin,
+    kSeekCurrent,
+    kSeekEnd
+};
 
-class Stream : public IAGSStream
+
+class Stream
 {
 public:
     Stream() = default;
     Stream(const String &path)
         : _path(path) {}
+    virtual ~Stream() = default;
 
     // Returns an optional path of a stream's source, such as a filepath;
     // primarily for diagnostic purposes
     const String &GetPath() const { return _path; }
+
+    //-----------------------------------------------------
+    // Stream interface
+    //-----------------------------------------------------
+
+    virtual bool        IsValid() const = 0;
     // Tells if the stream has errors
-    virtual bool HasErrors() const { return false; }
+    virtual bool        HasErrors() const { return false; }
+    virtual bool        EOS() const = 0;
+    virtual soff_t      GetLength() const = 0;
+    virtual soff_t      GetPosition() const = 0;
+    virtual bool        CanRead() const = 0;
+    virtual bool        CanWrite() const = 0;
+    virtual bool        CanSeek() const = 0;
+
+    virtual void        Close() = 0;
+
+    // Reads number of bytes in the provided buffer
+    virtual size_t      Read(void *buffer, size_t size) = 0;
+    // ReadByte conforms to fgetc behavior:
+    // - if stream is valid, then returns an *unsigned char* packed in the int
+    // - if EOS, then returns -1
+    virtual int32_t     ReadByte() = 0;
+    // Writes number of bytes from the provided buffer
+    virtual size_t      Write(const void *buffer, size_t size) = 0;
+    // WriteByte conforms to fputc behavior:
+    // - on success, returns the unsigned char packed in the int
+    // - on failure, returns -1
+    virtual int32_t     WriteByte(uint8_t b) = 0;
+
+    // Convenience methods for reading values of particular size
+    virtual int8_t      ReadInt8() = 0;
+    virtual int16_t     ReadInt16() = 0;
+    virtual int32_t     ReadInt32() = 0;
+    virtual int64_t     ReadInt64() = 0;
+    virtual size_t      ReadArray(void *buffer, size_t elem_size, size_t count) = 0;
+    virtual size_t      ReadArrayOfInt16(int16_t *buffer, size_t count) = 0;
+    virtual size_t      ReadArrayOfInt32(int32_t *buffer, size_t count) = 0;
+    virtual size_t      ReadArrayOfInt64(int64_t *buffer, size_t count) = 0;
+
+    // Convenience methods for writing values of particular size
+    virtual size_t      WriteInt8(int8_t val) = 0;;
+    virtual size_t      WriteInt16(int16_t val) = 0;
+    virtual size_t      WriteInt32(int32_t val) = 0;
+    virtual size_t      WriteInt64(int64_t val) = 0;
+    virtual size_t      WriteArray(const void *buffer, size_t elem_size, size_t count) = 0;
+    virtual size_t      WriteArrayOfInt16(const int16_t *buffer, size_t count) = 0;
+    virtual size_t      WriteArrayOfInt32(const int32_t *buffer, size_t count) = 0;
+    virtual size_t      WriteArrayOfInt64(const int64_t *buffer, size_t count) = 0;
+
+    virtual bool        Seek(soff_t offset, StreamSeek origin = kSeekCurrent) = 0;
+
     // Flush stream buffer to the underlying device
-    virtual bool Flush() = 0;
+    virtual bool        Flush() = 0;
 
     //-----------------------------------------------------
     // Helper methods
     //-----------------------------------------------------
-    bool ReadBool() override
+    bool ReadBool()
     {
         return ReadInt8() != 0;
     }
 
-    size_t WriteBool(bool val) override
+    size_t WriteBool(bool val)
     {
         return WriteInt8(val ? 1 : 0);
     }
 
     // Practically identical to Read() and Write(), these two helpers' only
     // meaning is to underline the purpose of data being (de)serialized
-    size_t ReadArrayOfInt8(int8_t *buffer, size_t count) override
+    size_t ReadArrayOfInt8(int8_t *buffer, size_t count)
     {
         return Read(buffer, count);
     }
-    size_t WriteArrayOfInt8(const int8_t *buffer, size_t count) override
+    size_t WriteArrayOfInt8(const int8_t *buffer, size_t count)
     {
         return Write(buffer, count);
     }
